Rename syscalls/write.c to syscalls/printf_putchar.c since that it what it actually implements. Also fix intermittent assert compilation error; see comments in flexpret_assert.h for explanation.
