+++ programs/lib/Makefrag-lib
@@ -8,7 +8,7 @@ else
 LIB_SOURCES += \
 	$(FP_LIB_DIR)/printf/src/printf/printf.c \
 	$(FP_LIB_DIR)/syscalls/syscalls.c \
-	$(FP_LIB_DIR)/syscalls/write.c \
+	$(FP_LIB_DIR)/syscalls/printf_putchar.c \
 	$(FP_LIB_DIR)/flexpret_thread.c \
 	$(FP_LIB_DIR)/flexpret_lock.c \
 	$(FP_LIB_DIR)/flexpret_cond.c \
